requirements.txt
numpy>=1.25
scipy>=1.11
scikit-learn>=1.4
PyYAML>=6.0
torch>=2.2
qiskit>=2.0
qiskit-aer>=0.17
qiskit-ibm-runtime>=0.43
